Move setting into their own page
